Low hanging IoActor refactor
