+++ engine/src/main/scala/cromwell/engine/io/IoActor.scala
@@ -1,7 +1,5 @@
 package cromwell.engine.io
 
-import java.io.IOException
-import java.net.{SocketException, SocketTimeoutException}
 import java.time.OffsetDateTime
 
 import akka.NotUsed
@@ -9,7 +7,6 @@ import akka.actor.{Actor, ActorLogging, ActorRef, Props, Timers}
 import akka.dispatch.ControlMessage
 import akka.stream._
 import akka.stream.scaladsl.{Flow, GraphDSL, Merge, Partition, Sink, Source, SourceQueueWithComplete}
-import com.google.cloud.storage.StorageException
 import com.typesafe.config.ConfigFactory
 import cromwell.core.Dispatcher.IoDispatcher
 import cromwell.core.actor.StreamActorHelper
@@ -18,12 +15,10 @@ import cromwell.core.io.{IoAck, IoCommand, Throttle}
 import cromwell.core.{Dispatcher, LoadConfig}
 import cromwell.engine.instrumentation.IoInstrumentation
 import cromwell.engine.io.IoActor._
-import cromwell.engine.io.gcs.GcsBatchFlow.BatchFailedException
 import cromwell.engine.io.gcs.{GcsBatchCommandContext, ParallelGcsBatchFlow}
 import cromwell.engine.io.nio.NioFlow
 import cromwell.filesystems.gcs.batch.GcsBatchIoCommand
 import cromwell.services.loadcontroller.LoadControllerService.{HighLoad, LoadMetric, NormalLoad}
-import javax.net.ssl.SSLException
 
 import scala.concurrent.ExecutionContext
 
@@ -180,69 +175,6 @@ object IoActor {
   case object BackPressureTimerResetKey
   case object BackPressureTimerResetAction extends ControlMessage
 
-  /**
-    * ATTENTION: Transient failures are retried *forever* 
-    * Be careful when adding error codes to this method.
-    * Currently only 429 (= quota exceeded are considered truly transient)
-    */
-  def isTransient(failure: Throwable): Boolean = failure match {
-    case gcs: StorageException => gcs.getCode == 429
-    case _ => false
-  }
-
-  def isGcs500(failure: Throwable): Boolean = {
-    val serverErrorPattern = ".*Could not read from gs.+500 Internal Server Error.*"
-    Option(failure.getMessage).exists(_.matches(serverErrorPattern))
-  }
-
-  def isGcs503(failure: Throwable): Boolean = {
-    val serverErrorPattern = ".*Could not read from gs.+503 Service Unavailable.*"
-    Option(failure.getMessage).exists(_.matches(serverErrorPattern))
-  }
-
-  def isGcs504(failure: Throwable): Boolean = {
-    val serverErrorPattern = ".*Could not read from gs.+504 Gateway Timeout.*"
-    Option(failure.getMessage).exists(_.matches(serverErrorPattern))
-  }
-
-  private val AdditionalRetryableHttpCodes = List(
-    // HTTP 410: Gone
-    // From Google doc (https://cloud.google.com/storage/docs/json_api/v1/status-codes):
-    // "You have attempted to use a resumable upload session that is no longer available.
-    // If the reported status code was not successful and you still wish to upload the file, you must start a new session."
-    410,
-    // Some 503 errors seem to yield "false" on the "isRetryable" method because they are not retried.
-    // The CloudStorage exception mechanism is not flawless yet (https://github.com/GoogleCloudPlatform/google-cloud-java/issues/1545)
-    // so that could be the cause.
-    // For now explicitly lists 503 as a retryable code here to work around that.
-    503
-  )
-
-  // Error messages not included in the list of built-in GCS retryable errors (com.google.cloud.storage.StorageException) but that we still want to retry
-  private val AdditionalRetryableErrorMessages = List(
-    "Connection closed prematurely"
-  ).map(_.toLowerCase)
-
-  /**
-    * Failures that are considered retryable.
-    * Retrying them should increase the "retry counter"
-    */
-  def isRetryable(failure: Throwable): Boolean = failure match {
-    case gcs: StorageException => gcs.isRetryable ||
-      isRetryable(gcs.getCause) ||
-      AdditionalRetryableHttpCodes.contains(gcs.getCode) ||
-      AdditionalRetryableErrorMessages.exists(gcs.getMessage.toLowerCase.contains)
-    case _: SSLException => true
-    case _: BatchFailedException => true
-    case _: SocketException => true
-    case _: SocketTimeoutException => true
-    case ioE: IOException if Option(ioE.getMessage).exists(_.contains("Error getting access token for service account")) => true
-    case ioE: IOException => isGcs500(ioE) || isGcs503(ioE) || isGcs504(ioE)
-    case other => isTransient(other)
-  }
-
-  def isFatal(failure: Throwable): Boolean = !isRetryable(failure)
-  
   def props(queueSize: Int,
             nioParallelism: Int,
             gcsParallelism: Int,
