+++ engine/src/main/scala/cromwell/engine/io/gcs/GcsBatchFlow.scala
@@ -16,7 +16,7 @@ import cromwell.cloudsupport.gcp.gcs.GcsStorage
 import cromwell.engine.io.IoActor._
 import cromwell.engine.io.IoAttempts.EnhancedCromwellIoException
 import cromwell.engine.io.gcs.GcsBatchFlow.{BatchFailedException, _}
-import cromwell.engine.io.{IoAttempts, IoCommandContext}
+import cromwell.engine.io.{IoAttempts, IoCommandContext, RetryableRequestSupport}
 import mouse.boolean._
 
 import scala.concurrent.duration._
@@ -42,7 +42,7 @@ object GcsBatchFlow {
 }
 
 class GcsBatchFlow(batchSize: Int, scheduler: Scheduler, onRetry: IoCommandContext[_] => Throwable => Unit, applicationName: String)
-                  (implicit ec: ExecutionContext) extends StrictLogging {
+                  (implicit ec: ExecutionContext) extends StrictLogging with RetryableRequestSupport {
 
   // Does not carry any authentication, assumes all underlying requests are properly authenticated
   private val httpRequestInitializer = new HttpRequestInitializer {
@@ -135,11 +135,11 @@ class GcsBatchFlow(batchSize: Int, scheduler: Scheduler, onRetry: IoCommandConte
     // Otherwise fail with the original exception
     Try(batchRequest.execute()) match {
       case Failure(failure: IOException) =>
-        logger.info(s"Failed to execute GCS Batch request. Failed request belonged to batch " +
+        logger.info(s"Failed to execute GCS Batch request. Failed request belonged to batch of size ${batchCommandNamesList.size} containing commands: " +
           s"${batchCommandNamesList.mkString("\n")}.", failure.toPrettyElidedString(limit = 1000))
         failAllPromisesWith(BatchFailedException(failure))
       case Failure(failure) =>
-        logger.info(s"Failed to execute GCS Batch request. Failed request belonged to batch " +
+        logger.info(s"Failed to execute GCS Batch request. Failed request belonged to batch of size ${batchCommandNamesList.size} containing commands: " +
           s"${batchCommandNamesList.mkString("\n")}.", failure.toPrettyElidedString(limit = 1000))
         failAllPromisesWith(failure)
       case _ =>
