+++ engine/src/test/scala/cromwell/engine/io/IoActorSpec.scala
@@ -20,6 +20,8 @@ import scala.concurrent.ExecutionContext
 import scala.concurrent.duration._
 import scala.language.postfixOps
 
+class ConcreteRetry extends RetryableRequestSupport
+
 class IoActorSpec extends TestKitSuite with AnyFlatSpecLike with Matchers with ImplicitSender {
   behavior of "IoActor"
   
@@ -260,8 +262,8 @@ class IoActorSpec extends TestKitSuite with AnyFlatSpecLike with Matchers with I
       new IOException("Some other text. Could not read from gs://fc-secure-<snip>/JointGenotyping/<snip>/call-HardFilterAndMakeSitesOnlyVcf/shard-4688/rc: 504 Gateway Timeout"),
     )
 
-    retryables foreach { IoActor.isRetryable(_) shouldBe true }
-    retryables foreach { IoActor.isFatal(_) shouldBe false }
+    retryables foreach { new ConcreteRetry().isRetryable(_) shouldBe true }
+    retryables foreach { new ConcreteRetry().isFatal(_) shouldBe false }
   }
 
   it should "have correct non-retryable exceptions" in {
@@ -274,7 +276,7 @@ class IoActorSpec extends TestKitSuite with AnyFlatSpecLike with Matchers with I
       new IOException("Could not read from gs://fc-secure-<snip>/JointGenotyping/<snip>/call-HardFilterAndMakeSitesOnlyVcf/shard-500/rc: 404 File Not Found")
     )
 
-    nonRetryables foreach {IoActor.isRetryable(_) shouldBe false}
-    nonRetryables foreach {IoActor.isFatal(_) shouldBe true}
+    nonRetryables foreach { new ConcreteRetry().isRetryable(_) shouldBe false }
+    nonRetryables foreach { new ConcreteRetry().isFatal(_) shouldBe true }
   }
 }
